Finance/requirements.txt
appdirs==1.4.3
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
DateTime==4.3
deprecation==2.0.7
eikon==1.0.1
graphviz==0.13
h5py==2.10.0
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
mock==3.0.5
multitasking==0.0.9
numexpr==2.7.0
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
patsy==0.5.1
pyparsing==2.4.2
pyro-api==0.1.1
pyro-ppl==0.4.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scipy==1.3.1
six==1.12.0
statsmodels==0.10.2
tables==3.5.2
toolz==0.10.0
torch==1.2.0
torchviz==0.0.1
tqdm==4.36.1
urllib3==1.25.6
websocket-client==0.56.0
yfinance==0.1.45
zope.interface==4.6.0
